Refactor the absolute path solution
Turns out utils.join_path() is not very smart about URLs, it will
happily concatenate the cwd with, for example, a YouTube URL, so we need
to check if path is a URL first. The "normalize-path" in the upcoming
mpv v0.39 should handle this case well, among with many other cases we
care about that the current utils.join_path() solution doesn't handle.